refactored the code, made it a lot easier to add new menu items to the paint menu, by making an interface and abstraction
