+++ src/main/java/Interfaces/MenuItemProducer.java
@@ -1,0 +1,11 @@
+package Interfaces;
+
+import javax.swing.*;
+
+public interface MenuItemProducer {
+    /**
+     * Method that produces a JPanel representing the menu item.
+     * @return The JPanel representing this menu item.
+     */
+    JPanel createMenuPanel();
+}
